DEV 002 visualizer mock data
